+++ packages/router/src/finish_response.ts
@@ -4,17 +4,17 @@ import {
   Response,
   RedirectLocation,
   SettableResponseProperties,
-  ResolveResults
+  ResolveResults,
+  IntrinsicResponse,
+  ResponseFn
 } from "@curi/types";
 
-import { RealMatch } from "./match_location";
-
 function create_redirect(
-  redirect_to: any,
+  redirect: any,
   interactions: Interactions,
   history: History
 ): RedirectLocation {
-  const { name, params, query, hash, state } = redirect_to;
+  const { name, params, query, hash, state } = redirect;
   const pathname = interactions.pathname(name, params);
   return {
     name,
@@ -27,22 +27,41 @@ function create_redirect(
   };
 }
 
+const valid_properties: {
+  [key in keyof SettableResponseProperties]: boolean
+} = {
+  status: true,
+  error: true,
+  body: true,
+  data: true,
+  title: true,
+  redirect: true
+};
+
+function valid_response_property(
+  property: string
+): property is keyof SettableResponseProperties {
+  if (process.env.NODE_ENV !== "production") {
+    if (!valid_properties.hasOwnProperty(property)) {
+      console.warn(`"${property}" is not a valid response property. The valid properties are:
+
+status, error, body, data, title, redirect`);
+    }
+  }
+  return valid_properties.hasOwnProperty(property);
+}
+
 export default function finish_response(
-  route_match: RealMatch,
+  create_response: ResponseFn,
+  match: IntrinsicResponse,
   interactions: Interactions,
   resolved_results: ResolveResults | null,
   history: History,
   external: any
 ): Response {
-  const { route, match } = route_match;
-  const response: Response = match;
-  if (!route.response) {
-    return response;
-  }
-
   const { resolved = null, error = null } = resolved_results || {};
 
-  const response_modifiers = route.response({
+  const response_modifiers = create_response({
     resolved,
     error,
     match,
@@ -50,27 +69,26 @@ export default function finish_response(
   });
 
   if (!response_modifiers) {
-    return response;
+    return match;
   }
 
-  const settable_properties: Array<
-    keyof Response & keyof SettableResponseProperties
-  > = ["status", "error", "body", "data", "title", "redirect_to"];
-
   // only merge the valid properties onto the response
-  settable_properties.forEach(p => {
-    if (response_modifiers.hasOwnProperty(p)) {
-      if (p === "redirect_to") {
-        // special case
-        response[p] = create_redirect(
-          response_modifiers[p],
-          interactions,
-          history
-        );
-      } else {
-        response[p] = response_modifiers[p];
+  return Object.keys(response_modifiers).reduce(
+    (acc, key) => {
+      if (valid_response_property(key)) {
+        if (key === "redirect") {
+          // special case
+          acc[key] = create_redirect(
+            response_modifiers[key],
+            interactions,
+            history
+          );
+        } else {
+          acc[key] = response_modifiers[key];
+        }
       }
-    }
-  });
-  return response;
+      return acc;
+    },
+    match as Response
+  );
 }