+++ packages/router/types/finish_response.d.ts
@@ -1,4 +1,3 @@
 import { History } from "@hickory/root";
-import { Interactions, Response, ResolveResults } from "@curi/types";
-import { RealMatch } from "./match_location";
-export default function finish_response(route_match: RealMatch, interactions: Interactions, resolved_results: ResolveResults | null, history: History, external: any): Response;
+import { Interactions, Response, ResolveResults, IntrinsicResponse, ResponseFn } from "@curi/types";
+export default function finish_response(create_response: ResponseFn, match: IntrinsicResponse, interactions: Interactions, resolved_results: ResolveResults | null, history: History, external: any): Response;