+++ packages/types/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* `redirect`, not `redirect_to`.
+
 ## 2.0.0-alpha.0
 
 * Move common types from `@curi/router` to this package.